Derive send and subscribe zones in one place
ZoneRules owns the single participant-type derivation, producing both
the send zones and the subscribable zones. StompAuthorizer composes it,
keeping only the connection-scoped concerns (temporary grants and
replyToId reply scoping), and StompAuthorizerFactory is gone — the
connection handler constructs the authorizer directly. The zone matrix
test drives StompAuthorizer through the same construction path.
