+++ website/content/2.platform/6.defense-in-depth.md
@@ -16,7 +16,7 @@ Two principles run through all of it:
 
 ## Zone Authorization
 
-Every routable address carries its zone in the CRI itself, so the isolation boundary travels with the message and is validated wherever the message goes. Which zones a participant may address is derived from the participant type in one place — the gateway's `ZoneSendRules` — and every enforcement point uses that same derivation. The full participant × zone matrix is in the [CRI format reference](/platform/reference/cri-format).
+Every routable address carries its zone in the CRI itself, so the isolation boundary travels with the message and is validated wherever the message goes. Which zones a participant may address is derived from the participant type in one place — the gateway's `ZoneRules` — and every enforcement point uses that same derivation. The full participant × zone matrix is in the [CRI format reference](/platform/reference/cri-format).
 
 The derivation defends its own inputs:
 
@@ -37,7 +37,7 @@ Every frame on a connection is checked against the participant's zones — authe
 An MCP `tools/call` crosses two independent authorization layers:
 
 1. **Resolution is scoped.** The tool lookup queries Elasticsearch with the caller's zone visibility filter — a tool outside the participant's zones is not found at all, exactly as if it did not exist.
-2. **Dispatch is re-authorized.** Before the call is sent, the invoker re-checks the resolved CRI against the same `ZoneSendRules` the STOMP path enforces. If the stored directory data or the visibility query were ever wrong, the dispatch is refused, the CRIs are logged as a server fault, and the caller sees an unknown tool.
+2. **Dispatch is re-authorized.** Before the call is sent, the invoker re-checks the resolved CRI against the same `ZoneRules` the STOMP path enforces. If the stored directory data or the visibility query were ever wrong, the dispatch is refused, the CRIs are logged as a server fault, and the caller sees an unknown tool.
 
 The surrounding surface is hardened the same way:
